chore(package): package functions individually - EUBFR-35 (#18)
* chore(package): package functions individually - EUBFR-35

* Update existing services
