+++ services/ingestion/etl/budg/csv/src/events/onParseCSV.js
@@ -1,16 +1,15 @@
-/* eslint-disable import/prefer-default-export, no-console */
 import path from 'path';
 import stream from 'stream';
 import AWS from 'aws-sdk'; // eslint-disable-line import/no-extraneous-dependencies
 import parse from 'csv-parse';
 import transform from 'stream-transform';
 
-import transformRecord from './transform';
+import transformRecord from '../lib/transform';
 
 // Destination bucket
 const { BUCKET } = process.env;
 
-export const parseCsv = (event, context, callback) => {
+export const handler = (event, context, callback) => {
   /*
    * Some checks here before going any further
    */
@@ -87,3 +86,5 @@ export const parseCsv = (event, context, callback) => {
     .pipe(transformer)
     .pipe(uploadFromStream());
 };
+
+export default handler;
